refactor: Create a OilOingMain
