+++ presentation/src/main/java/oil/oing/nearby/presentation/ui/OilOingMain.kt
@@ -1,4 +1,70 @@
 package oil.oing.nearby.presentation.ui
 
-class OilOingMain {
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Button
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import com.naver.maps.map.compose.ExperimentalNaverMapApi
+import com.naver.maps.map.compose.MapProperties
+import com.naver.maps.map.compose.MapUiSettings
+import com.naver.maps.map.compose.NaverMap
+import oil.oing.nearby.presentation.theme.OilOingTheme
+
+@Composable
+fun OilOingMain() {
+    OilOingTheme {
+        Surface(modifier = Modifier.fillMaxSize(), color = OilOingTheme.background.color) {
+            MainScreen()
+        }
+    }
+}
+
+@OptIn(ExperimentalNaverMapApi::class)
+@Composable
+fun MainScreen() {
+    var mapProperties by remember {
+        mutableStateOf(
+            MapProperties(maxZoom = 17.0, minZoom = 10.0)
+        )
+    }
+    var mapUiSettings by remember {
+        mutableStateOf(
+            MapUiSettings(isLocationButtonEnabled = false)
+        )
+    }
+    Box(Modifier.fillMaxSize()) {
+        NaverMap(properties = mapProperties, uiSettings = mapUiSettings)
+        Column {
+            Button(onClick = {
+                mapProperties = mapProperties.copy(
+                    isBuildingLayerGroupEnabled = !mapProperties.isBuildingLayerGroupEnabled
+                )
+            }) {
+                Text(text = "Toggle isBuildingLayerGroupEnabled")
+            }
+            Button(onClick = {
+                mapUiSettings = mapUiSettings.copy(
+                    isLocationButtonEnabled = !mapUiSettings.isLocationButtonEnabled
+                )
+            }) {
+                Text(text = "Toggle isLocationButtonEnabled")
+            }
+        }
+    }
+}
+
+@Preview
+@Composable
+private fun MainScreenPreview(
+) {
+    MainScreen()
 }
